requirements.txt
django~=3.1.1
django-crispy-forms
psycopg2-binary
psycopg2
django-postgres-copy
pandas~=1.1.3
djangorestframework~=3.12.1
django-crontab
django-filter~=2.4.0
PyJWT==1.7.1
hyper